pyproject.toml
[tool.poetry]
package-mode = false
name = "discord-ai-agent-npc"
version = "0.1.0"
description = ""
authors = ["Abiorh001 <abiolaadedayo1993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
discord-py = "^2.4.0"
python-decouple = "^3.8"
langchain-openai = "^0.3.2"
langgraph = "^0.2.67"
langchain = "^0.3.15"
langchain-community = "^0.3.15"
qdrant-client = "^1.13.2"
llama-index = "^0.12.13"
python-dotenv = "^1.0.1"
cohere = "^5.13.11"
llama-index-postprocessor-cohere-rerank = "^0.3.0"
llama-index-retrievers-bm25 = "^0.5.2"
llama-index-vector-stores-pinecone = "^0.4.3"
llama-index-vector-stores-qdrant = "^0.4.3"
llama-index-readers-file = "^0.4.4"
feedparser = "^6.0.11"
textblob = "^0.19.0"
pydantic = "^2.10.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
